requirements.txt
langgraph>=0.0.10
langchain>=0.1.0
langchain-openai>=0.0.2
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.2
pydantic>=2.5.2
rich>=13.7.0
pyinstaller>=6.3.0
pillow>=10.2.0 